[ENH] Add command line options and auto execution
Added command line parameter parsing and auto execution

Add the option to use the application from the command line or pass
the parameter to the gui. You can also specify if the conversion
will be started automatically, after completion the application will
be closed. Additional bugfix to add ignoreErrors to the settings.